% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getAnnualObsTable.R
\name{getAnnualObsTable}
\alias{getAnnualObsTable}
\title{Get the section of the csv file that comprises the Annual Observations.}
\usage{
getAnnualObsTable(df)
}
\arguments{
\item{df}{The imported dataframe.}
}
\value{
A dataframe of annual observations from the csv file.
}
\description{
Get the section of the csv file that comprises the Annual Observations.
}

