% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getDomain.R
\name{getDomain}
\alias{getDomain}
\title{Get the domain.}
\usage{
getDomain(file)
}
\arguments{
\item{file}{The imported csv file.}
}
\value{
The domain from the csv file.
}
\description{
Get the domain.
}

