% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getPhenoObsTable.R
\name{getPhenoObsTable}
\alias{getPhenoObsTable}
\title{Get the section of the csv file that comprises the Phenology Observations.}
\usage{
getPhenoObsTable(df)
}
\arguments{
\item{df}{The imported dataframe.}
}
\value{
A dataframe of phenological observations from the csv file.
}
\description{
Get the section of the csv file that comprises the Phenology Observations.
}

