% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getPlantSamplesTable.R
\name{getPlantSamplesTable}
\alias{getPlantSamplesTable}
\title{Get the section of the csv file that comprises the Plant Samples data.}
\usage{
getPlantSamplesTable(df)
}
\arguments{
\item{df}{The imported dataframe.}
}
\value{
A dataframe of plant sample observations from the csv file.
}
\description{
Get the section of the csv file that comprises the Plant Samples data.
}

