% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getProtocol.R
\name{getProtocol}
\alias{getProtocol}
\title{Get the protocol code}
\usage{
getProtocol(file)
}
\arguments{
\item{file}{The imported csv file.}
}
\value{
The protocol code from the csv file.
}
\description{
Get the protocol code
}

