% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mergeTables.R
\name{mergeTables}
\alias{mergeTables}
\title{Merge the two dataframes that were retrieved from the csv file.}
\usage{
mergeTables(plantsamples, phenoobs, anntable)
}
\arguments{
\item{plantsamples}{The dataframe of plant samples.}

\item{phenoobs}{The dataframe of phenology observations.}
}
\value{
A dataframe of merged data from the csv file.
}
\description{
Merge the two dataframes that were retrieved from the csv file.
}

