\name{neonPhenoSheets}
\alias{neonPhenoSheets-package}
\alias{neonPhenoSheets}
\docType{package}
\title{
Imports past phenology bout data into new data sheet
}
\description{
This package reads the .csv file output from the National Ecological Observatory Network's (NEON) data entry application for phenology, pulls out information relevant for the next bout, and creates a new data sheet. Plant individuals that are marked to be dropped in the previous bout are not included in the next bout data sheet. 
}
\details{
\tabular{ll}{
Package: \tab neonPhenoSheets\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-03-02\cr
License: \tab GPL (>= 2)\cr
}
This package is to be used internally by NEON to generate new data sheets as needed.  
}
\author{
Christine M. Laney

Maintainer: Christine Laney <claney@neoninc.org>
}
\references{
None
}
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
