% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parsePhenoCSV.R
\name{parsePhenoCSV}
\alias{parsePhenoCSV}
\title{Parse the phenology csv file.}
\usage{
parsePhenoCSV(file)
}
\arguments{
\item{file}{The imported csv file.}
}
\value{
A new dataframe built from information parsed from the csv file.
}
\description{
Parse the phenology csv file.
}

